2022/12/test_aoc202212.py
"""Tests for AoC 12, 2022."""

# Standard library imports
import pathlib

# Third party imports
import aoc202212
import pytest

PUZZLE_DIR = pathlib.Path(__file__).parent


@pytest.fixture
def example1():
    puzzle_input = (PUZZLE_DIR / "example1.txt").read_text().rstrip()
    return aoc202212.parse_data(puzzle_input)


@pytest.fixture
def example2():
    puzzle_input = (PUZZLE_DIR / "example2.txt").read_text().rstrip()
    return aoc202212.parse_data(puzzle_input)


def test_parse_example1(example1):
    """Test that input is parsed properly."""
    assert example1 == [['S','a','b','q','p','o','n','m'],
                        ['a','b','c','r','y','x','x','l'],
                        ['a','c','c','s','z','E','x','k'],
                        ['a','c','c','t','u','v','w','j'],
                        ['a','b','d','e','f','g','h','i']]


def test_part1_example1(example1):
    """Test part 1 on example input."""
    assert aoc202212.part1(example1) == 31


def test_part2_example1(example1):
    """Test part 2 on example input."""
    assert aoc202212.part2(example1) == 29
